feat(NumberAdder) finish

+++ week-03/day-04/src/Counter.java
@@ -1,10 +1,10 @@
 public class Counter {
   public static void main(String[] args) {
-  int n = 510;
+  long n = 510;
   recursiveCounter(n);
   }
 
-  public static void recursiveCounter (int n){
+  public static void recursiveCounter (long n){
     if ( n <= 0){
       System.out.println(0);
       return;
